dso: preload needed shared objects manually
The symbol execve() from the libc resolves the needed shared objects
before the given shared object executable file is actually executed.

It might result in loading the shared objects from the host instead of
the ones from the chroot'ed environment.

This preloads the needed shared objects manually, before the executable
file is execve()'ed.

Note: It implements the dynamic loader searches as defined in ld.so(8).