dso: look up in both lib and exec RPATH directories
TL;DR; The man-pages ld.so(8)[1] and dlopen(3)[2] are both incompletes
about the shared object lookup algorithm. Indeed, both shared objects
(the library **AND** the executable file) RPATH tags are used to search
for the shared object (i.e. the loading shared object too; not only the
executable file).

However, both RPATH tags are unused if the loading shared object (i.e
the dlopen(3)'ed library) if its RUNPATH is set.

Additionnaly, the RPATH tag of the executable file is unused if its
RUNPATH is set (still if the dlopen(3)'ed library RPATH tag is set and
RUNPATH tag is unset).

This fixes the shared object lookup algorithm after digging into the
glibc's ld.so(8) code to look up in both library and executable file
RPATH directories.

Note: It is unclear if the library path lookup is chained; thus, it may
not be implemented correctly; however, it is good enough enough for now.

The long story starts in the Qt blog post[3], and its author thread in
the man-page mailing list[4], and the RPATH issue in Debian wiki[5].

This rises the good lookup algorithm:

	Unless loading object has RUNPATH:
	  RPATH of the loading object, then the RPATH of its loader
	  (unless it has a RUNPATH), ..., until the end of the chain,
	  which is either the executable or an object loaded by dlopen
	  Unless executable has RUNPATH:
	    RPATH of the executable
	LD_LIBRARY_PATH
	RUNPATH of the loading object
	ld.so.cache
	default dirs

Note: The loading object is the shared object to load, i.e. given to
dlopen(3).

The first point of dlopen(3) is incomplete and even incorect:

	•  (ELF only) If the calling object (i.e., the shared library or
	executable from which dlopen() is called) contains a DT_RPATH
	tag, and does not contain a DT_RUNPATH tag, then the directories
	listed in the DT_RPATH tag are searched.

It should be corrected by:

	•  (ELF only) If the calling object (i.e., the shared library
	from which dlopen() is called) contains a DT_RPATH tag, and does
	not contain a DT_RUNPATH tag, then the directories listed in the
	DT_RPATH tag are searched. If the executable object contains a
	DT_RPATH tag, and does not contain a DT_RUNPATH tag, then the
	directories listed in the DT_RPATH tag are searched.

But it remains incomplete if it turns out the needed shared objects are
of chain opened.

	If the object specified by filename has dependencies on other
	shared objects, then these are also automatically loaded by the
	dynamic linker using the same rules. (This process may occur
	recursively, if those objects in turn have dependencies, and so
	on.)

This takes the glibc's dynamic loader comments from elf/dl-load.c[6] and
elf/get-dynamic-info.h[7] for later investigations.

Note: The code is not taken from musl as the libc and dl.so are the same
object.

[1]: https://linux.die.net/man/8/ld.so
[2]: https://linux.die.net/man/3/dlopen
[3]: https://www.qt.io/blog/2011/10/28/rpath-and-runpath
[4]: https://www.spinics.net/lists/linux-man/msg02291.html
[5]: https://wiki.debian.org/RpathIssue
[6]: https://sourceware.org/git/?p=glibc.git;a=blob;f=elf/dl-load.c;hb=glibc-2.37#l2023
[7]: https://sourceware.org/git/?p=glibc.git;a=blob;f=elf/get-dynamic-info.h;hb=glibc-2.37#l185