exec.sh: detect if -r is already given to ldconfig
The ldconfig might be run with option -r; therefore the option -r is
given twice to the rewritten command: the first one with the given root
path (not rewritten the chroot'ed environment), the second one with the
prepend $IAMROOT_ROOT path.

This rewrites the ldconfig command to prepend the $IAMROOT_ROOT path to
the option -r if set, and to append the option -r $IAMROOT_ROOT only if
the option -r is unset.