exec.sh: detect if -f is given to ldconfig
The ldconfig might be run with option -f; therefore the option -f has to
be reused instead of the default /etc/ld.so.conf.

This rewrites the ldconfig command to sed the conffile of the option -f
if set, or to sed the default conffile /etc/ld.so.conf if the option -f
is unset.