dso: implement ldd ld.so's requirements internally
TL;DR; This implements the ld.so's --verify and LD_TRACE_LOADER_OBJECTS
features internally to run ldd; this might be a temporary hack before a
better solution is found or before a true dynamic loader is implemented
or before a dynamic loader wrapper binary is implemented, who knows?

That solution is prefered for now as some part of the logic is already
there.

The long story then...

The command ldd is a GNU libc bash script wrapper on top of the dynamic
loader ld.so, which is naturally a statically linked program. Therefore,
it is impossible to preload the iamroot library to do path_resolution
for the programs given in parameters, and it opens the programs from the
host system instead of the ones from the chroot'ed environment.

It is architecture specific as it hardcodes the list of the rtld
filenames supported for the given architecture[1].

Here are some examples:

	x86_64: RTLDLIST="/usr/lib/ld-linux.so.2 /usr/lib64/ld-linux-x86-64.so.2 /usr/libx32/ld-linux-x32.so.2"
	i386: RTLDLIST=/usr/lib/ld-linux.so.2
	aarch64: RTLDLIST=/usr/lib/ld-linux-aarch64.so.1
	armhf: RTLDLIST=/usr/lib/ld-linux-armhf.so.3
	riscv64: RTLDLIST="/usr/lib/ld-linux-riscv64-lp64.so.1 /usr/lib/ld-linux-riscv64-lp64d.so.1 /usr/lib/ld-linux-riscv32-ilp32.so.1 /usr/lib/ld-linux-riscv32-ilp32d.so.1"

First, the ldd script verifies the first matching rtld that supports the
given program the via the command below[2]:

	ld.so --verify PROGRAM

Where ld.so is a rtld listed in $RTLDLIST[1].

It returns:

 - 2 if the PROGRAM is not dynamically linked
 - 0 if the PROGRAM is dynamically linked and if this dynamic linker can
   handle it
 - 1 otherwise if the PROGRAM is a non(-supported) ELF binary or it is
   no binary at all

See the examples below:

	gportay@archlinux ~/src/iamroot $ /lib64/ld-linux-x86-64.so.2 --verify /usr/bin/bash; echo $?
	0
	gportay@archlinux ~/src/iamroot $ /lib64/ld-linux-x86-64.so.2 --verify /usr/lib/libc.so.6; echo $?
	0
	gportay@archlinux ~/src/iamroot $ /lib64/ld-linux-x86-64.so.2 --verify /lib64/ld-linux-x86-64.so.2; echo $?
	2
	gportay@archlinux ~/src/iamroot $ /lib64/ld-linux-x86-64.so.2 --verify /; echo $?
	1
	gportay@archlinux ~/src/iamroot $ /lib/ld-linux.so.2 --verify /usr/lib64/libc.so.6; echo $?
	1

Then, the ldd script traces the PROGRAM (if the command --verify has
returned 0 or 2[3]) via the command below[4]:

	LD_TRACE_LOADED_OBJECTS=1 ld.so PROGRAM

Where ld.so is the first supported rtld listed in $RTLDLIST[1], and
where LD_TRACE_LOADED_OBJECTS is set[5] to 1 to cause the PROGRAM to
list its dynamic dependencies, instead of running normally.

Alternative solutions was also considered.

For example, intercepting the command ldd, restarting the iamroot
environment from the script exec.sh (i.e. handling if the program is an
interpreter script and its argument if any, preloading the libraries and
set the library path of the chroot'ed environment...); in short, it is
re-implementing the magic in shell.

However, the call to the dynamic loader (rtld in ldd) needs to be
tweaked.

The $IAMROOT_ROOT directory must be added to:
 - the command-line (to the program path and the shared object
   arguments)
 - the optional arguments --preload and --library-path (to every shared
   object and directories)
 - the environment variables LD_PRELOAD and LD_LIBRARY_PATH (to every
   shared object and directories)

The whorse part is to keep track of original values of the original
command line to strip them later if needed.

The dynamic loader output has to be tweaked to strip off the path to the
chrood'ed environment.

Therefore, the call to the dynamic load should either be intercepted
and/or rewritten from the library.

The two notes below shows some results of calling the dynamic loader of
a foreign rootfs from the host.

Note: The --verify command from a standard shell environment:

	gportay@archlinux ~/src/iamroot $ aarch64-archlinuxarm-rootfs/lib/ld-linux-aarch64.so.1 --verify aarch64-archlinuxarm-rootfs/usr/bin/bash; echo $?
	0
	gportay@archlinux ~/src/iamroot $ aarch64-archlinuxarm-rootfs/lib/ld-linux-aarch64.so.1 --verify aarch64-archlinuxarm-rootfs/; echo $?
	1
	gportay@archlinux ~/src/iamroot $ aarch64-archlinuxarm-rootfs/lib/ld-linux-aarch64.so.1 --verify aarch64-archlinuxarm-rootfs/lib/ld-linux-aarch64.so.1; echo $?
	2

Note 2: The command LD_TRACE_LOADED_OBJECTS=1 from a standard shell environment:

	gportay@archlinux ~/src/iamroot $ LD_TRACE_LOADED_OBJECTS=1 aarch64-archlinuxarm-rootfs/lib/ld-linux-aarch64.so.1 aarch64-archlinuxarm-rootfs/usr/bin/bash; echo $?
	/lib/ld-linux-aarch64.so.1 => aarch64-archlinuxarm-rootfs/lib/ld-linux-aarch64.so.1 (0x0000005500000000)
	libreadline.so.8 => not found
	libc.so.6 => not found
	0

	gportay@archlinux ~/src/iamroot $ LD_TRACE_LOADED_OBJECTS=1 aarch64-archlinuxarm-rootfs/lib/ld-linux-aarch64.so.1 --library-path aarch64-archlinuxarm-rootfs/lib  aarch64-archlinuxarm-rootfs/usr/bin/bash; echo $?
	libreadline.so.8 => aarch64-archlinuxarm-rootfs/lib/libreadline.so.8 (0x0000005502980000)
	libc.so.6 => aarch64-archlinuxarm-rootfs/lib/libc.so.6 (0x00000055029f0000)
	/lib/ld-linux-aarch64.so.1 => aarch64-archlinuxarm-rootfs/lib/ld-linux-aarch64.so.1 (0x0000005500000000)
	libncursesw.so.6 => aarch64-archlinuxarm-rootfs/lib/libncursesw.so.6 (0x0000005502bc0000)
	0

The dynamic loader wrapper appears to be an interresting solution; it
would set the libraries to preload, the library path to look up, the
alternative root... and it might be reused in ish.

	$ /usr/lib/iamroot/ld.so [--root /path/to/rootfs] [--library /path/to/library] [--library-path library:path] [OPTIONS] [PROGRAM [ARGUMENTS]]

This justs throw a stone into the pond.

[1]: https://sourceware.org/git/?p=glibc.git;a=blob;f=elf/ldd.bash.in;hb=glibc-2.37#l29
[2]: https://sourceware.org/git/?p=glibc.git;a=blob;f=elf/ldd.bash.in;hb=glibc-2.37#l159
[3]: https://sourceware.org/git/?p=glibc.git;a=blob;f=elf/ldd.bash.in;hb=glibc-2.37#l175
[4]: https://sourceware.org/git/?p=glibc.git;a=blob;f=elf/ldd.bash.in;hb=glibc-2.37#l116
[5]: https://sourceware.org/git/?p=glibc.git;a=blob;f=elf/ldd.bash.in;hb=glibc-2.37#l103