dlopen, dlmopen, dso: add internals __{dl,ldso}_access()
This introduces the internal functions __dl_access() and __ldso_access()
to resolve the shared object as described by ld.so(8).

Note: The function __dl_access() is an helper on top of the function
__dlso_access() used by the functions dlopen() and dlmopen().

Important: It is not used to resolved the needed shared objects for now;
it intends to deprecated __dl_lib_path() in the future.