Fix "exec.sh: detect if -r is already given to ldconfig"
The commit dfdcabff56c5863933b2009be2da68039af35c41 is buggy.

First, it intends to detect if the option -r is given through the
command line. It detects if the previous argument in $prev is -r, but
the variable is not updated in the for loop.

Second, it modifies the arguments in the variable $args, by prepending
them by the root directory $IAMROOT_ROOT. It includes the positional
arguments (library, directory) and the optional arguments (i.e. the
options -nNvVX [-C cache] [-f conf] [-r root], -l [-v] and -p).
Hopefully, the variable $args is left unused.

Third, it does not resolve correctly the root directory $IAMROOT_ROOT if
the path is relative (i.e. not absolute).

This addresses the three issues mentioned above, by setting the variable
prev in the loop for, by predending the relevant arguments, by resolving
correctly the path using the function path_resolution() and by using the
variable args.

This fixes commit dfdcabff56c5863933b2009be2da68039af35c41.