mk*temp*: fix copy back of generated temporary filename
The function path_resolution() sanitizes the generated temporary path
which may shorten the template string; it may strip the leading ./, /.,
/...

Therefore, the copyback of the generated string could not rely on the
template string length as they may differ.

Ideally, it should copyback only the last 6 XXXXXX characters and takes
in account the suffix if any.

However, it is safe to copy the whole basename.

This fixes the copyback of the generated temporary filename to the
template by copying the basename only.

Fixes:

	Failure writing database .//etc/udev/hwdb.bin: No such file or directory