Fix "Add IAMROOT_VERSION"
The purpose of the environment variable IAMROOT_VERSION is to give an
indication to the script exec.sh about the API as it has changed accross
the version and it can be considered as unstable.

This restricts the set of the IAMROOT_VERSION to the script by moving it
to the function exec_sh().

Additionnaly, this fixes potential invalid reads due to the use of the
function setenv() that may change the global variable __environ.

This is pretty bad as the function __ldso() uses strings from the
environment in the additional argument array reserved to the dynamic
loader. See the commits 4603aa858b7cc086a94228b262592e1f00f9852b,
e07475f800b3e673a7864d4edf11ac1a4168bc20.

This fixes commit cdbcb0fc087883a83ebb335e18d5c067fd843ae0.