execve: workaround nodo environment functions to run exec.sh
For a reason that is still unknown, in the GNU C Library world, the
three environment functions putenv(), setenv() and unsetenv() do nothing
in the execve() context call.

The two global variables environ and __environ remain unchanged.

However, the function clearenv() clears the environ and sets the two
globals to NULL.

It is mandatory to reset the environment variable LD_PRELOAD before
calling the script exec.sh from the host.

This clears the environment, and copies a bare minimal environment to
run the script exec.sh.