execve: workaround nodo environment functions to run exec.sh
The library and the script exec.sh are host system binaries, it must be
run from the host system. Furthermore, the intepreter for the script
exec.sh is not necessarily part of the chroot environment.

The GNU C Library leaks symbols in the dynamically linked binaries.
Therefore, the binaries MUST load the same or any later version of the
libc. As a consequence, the exec.sh MUST be link at runtime against the
GNU libc from the host system, which implies the dynamic loader
environment variables LD_PRELOAD and LD_LIBRARY_PATH MUST be unset, as
they may be set to chroot environment libraries and paths since the
linux dynamic loader prior to 2.30 do not have support for option
--preload and the LD_PRELOAD is used instead.

For a reason that is still unknown, in the GNU C Library world, the
three environment functions putenv(), setenv() and unsetenv() do nothing
in the execve() context call. The two global variables environ and
__environ remain unchanged. However, the function clearenv() clears the
environ and sets the two globals to NULL.

It workarounds the following error as LD_PRELOAD is STILL set to preload
the GNU libc in version 2.29 in the chroot environment, but exec.sh is
run via the /usr/bin/env linked against the GNU libc in version 2.38 in
the host system:

	/usr/bin/env: /usr/lib64/libc-2.29.so: version `GLIBC_2.38' not found (required by /usr/bin/env)
	/usr/bin/env: /usr/lib64/libc-2.29.so: version `GLIBC_2.34' not found (required by /usr/bin/env)
	/usr/bin/env: /usr/lib64/libc-2.29.so: version `GLIBC_ABI_DT_RELR' not found (required by /usr/lib/libpthread.so.0)

This sets a bare minimal environment to run the script exec.sh instead
of using the __environ.