tests: fix broken link in rootfs
The catalog file is created and installed early on Linux; before the
library directories and symlinks are actually made.

It results in a broken symlink on GNU/Linux 64-bits: the symlink
rootfs/usr/lib/lib64 links to itself (-> lib64):

	$ ls tests/rootfs/usr/lib -l
	total 8
	lrwxrwxrwx 1 gportay gportay   5 Sep 27 20:08 lib64 -> lib64
	drwxr-xr-x 1 gportay gportay   6 Sep 27 20:08 nls
	-rw-r--r-- 1 gportay gportay 172 Sep 27 20:08 os-release

This fixes the broken link by creating and installing the catalog file
after the library directories and symlinks are made.