Add IAMROOT_PATH_RESOLUTION_AF_UNIX
The Unix familly socket is limited to a 108-character long path; the
NULL-terminating character is not included (ie. 107 + NULL).

Prepending the root directory to resolve the real in-chroot path may
exceed that limitation and causes a ENAMETOOLONG error.

This introduces the environment variable IAMROOT_PATH_RESOLUTION_AF_UNIX
to reduce the path resolution of an in-chroot Unix socket and tries to
not exceed the limitation of the path length.