exec.sh: unset dynamic loader environment variables
For an unknown reason, the LD_PRELOAD is not reset before the script
exec.sh is execve'ed in the GNU C Library world.

Tests show the three environment functions putenv(), setenv() and
unsetenv() do nothing in the execve() context call.

This ensures the dynamic loader environment variables are unset before
running any executable in the script exec.sh; this avoids the ignored
dynamic executables to loop for eternity.

Fixes:

	$ mountpoint rootfs
	(... loop forever and beyond!...)