Fix "Add __{,f,l}stat{,at}_time64 weak aliases" once again
The 64-bit time stat symbols were introduced with glibc-2.33 to prepare
the glibc for y2038[1] (aka Epochalypse, 03:14:07 UTC on 19 January
2038, i.e. 32-bit time overlow to Epoch, 00:00:00 UTC on 1 January
1970).

First for LFS[2] (Large File Support, i.e. 64-bit time definition on
64-bit offset), then for legacy ABI[3].

It appears these symbols are used on the 32-bit architectures x86 and
arm (eabi, hf); and they are not used on the 64-bit architectures
x86_64, aarch64 and riscv64 (lp64).

This removes the weak aliases and implements the functions if the
timesize is 32-bits.

This fixes commit b10f732dfecd5b318a9d427b9af1a5d309ab41fd.

[1]: https://sourceware.org/git/?p=glibc.git;a=commit;h=aa03f722f3b994aaf81e72a8904bf33196780930
[2]: https://sourceware.org/git/?p=glibc.git;a=commit;h=47f24c21ee38701ae275aa9e451f70fa3e77478c
[3]: https://www.gnu.org/software/gnulib/manual/html_node/Avoiding-the-year-2038-problem.html