ish: fix ISH_USER and let shell sets USER
This fixes commit 7093318b1c25d94587c9baecc32ee5f80cae2cab.

This fixes the setting of the variable ISH_USER by removing the setting
for the variable USER which causes the ISH_USER to get set to mismatch
value. The variable USER is set later by the shell instead.

Fixes:

	gportay@archlinux ~/src/iamroot $ ish <<<'echo ISH_USER=$ISH_USER'
	ISH_USER=root
	gportay@archlinux ~/src/iamroot $ ish <<<'echo ISH_UID=$ISH_UID'
	ISH_UID=1000
	gportay@archlinux ~/src/iamroot $ ish <<<'echo USER=$USER'
	USER=root