linux.mk: switch to ido and add IDOFLAGS variable
The commands used to build the rootfs are mostly run via sudo to switch
the user to the super user root.

Since ido has been introduced, it makes lots of more sense to use it
instead of ish.

Furthermore, some ido's features from ish will be dropped in a close
future; in particular:
 - preserve environment (-E, --preserve-env)
 - target user/group (--user, --group)

This switches to ido to run the commands used to create the rootfs; it
introduces the variable IDOFLAGS to customize the ido command-line for
the needs of specific distributions.