Support chroot'ing to a PowerPC64 GNU world
The architecture PowerPC64 (i.e. big endian) uses the elfv1 ABI while
the little endian uses the elfv2 ABI.

Note: The musl library uses elfv2 ABI for both PowerPC64 and PowerPC64
little endian architectures[1][2].

The project uses the pre-built toolchains provided by Bootlin[3] since
commit b00046e4bbf1a6849f081a5c398d16be0b52cda3 to compile the libraries
targetting a foreign world (i.e. Aarch64, ARM, ARMHF, PowerPC64, MIPS,
S390 on x86-64).

Bootlin uses Buildroot to make its toolchains[4] and it ships toolchains
with their default ABI:
 - PowerPC64 glibc: elfv1
 - PowerPC64 LE glibc: elfv2
 - PowerPC64 and PowerPC64 LE musl: elfv2[5]

ArchLinuxPOWER[6], the unofficial port of Archlinux to PowerPC, uses the
GNU C Library userland and ships prebuilt PowerPC64 and PowerPC64 LE
binary packages using both the ABI elfv2.

As a consequence, the Bootlin PowerPC64 POWER8 glibc toolchain[7] cannot
be used to make the PowerPC64 ArchPOWER root filesystem as the toolchain
uses ABI elfv1.

Unfortunatly, Buildroot do not have a Kconfig option to select the
PowerPC ABI.

However, the diff below allows to build a PowerPC64 toolchain with the
ABI elfv2:

	diff --git a/package/gcc/gcc.mk b/package/gcc/gcc.mk
	index 94935d79a2..d94b72c544 100644
	--- a/package/gcc/gcc.mk
	+++ b/package/gcc/gcc.mk
	@@ -270,6 +270,14 @@ ifeq ($(BR2_powerpc)$(BR2_powerpc64),y)
	 HOST_GCC_COMMON_CONF_OPTS += --enable-secureplt
	 endif

	+# PowerPC64 big endian by default uses the elfv1 ABI, and PowerPC64 little
	+# endian by default uses the elfv2 ABI. However, ArchLinuxPOWER has decided to
	+# use the elfv2 ABI for both, so we force the elfv2 ABI for PowerPC64 big
	+# endian when the selected C library is glibc.
	+ifeq ($(BR2_TOOLCHAIN_USES_GLIBC)$(BR2_powerpc64),yy)
	+HOST_GCC_COMMON_CONF_OPTS += --with-abi=elfv2
	+endif
	+
	 # PowerPC64 big endian by default uses the elfv1 ABI, and PowerPC 64
	 # little endian by default uses the elfv2 ABI. However, musl has
	 # decided to use the elfv2 ABI for both, so we force the elfv2 ABI for

Follow the commands below to build the toolchain and to install it to the
host system.

First, clone the repository and patch the sources with the trivial
change above:

	$ git clone https://gitlab.com/buildroot.org/buildroot.git
	$ cd buildroot

Then, configure using QEMU PowerPC64 LE (POWER8) and update the target
architecture to PowerPC64 (big endian):

	$ make qemu_ppc64le_powernv8_defconfig
	$ make menuconfig
	(... Go to: "Target options  --->"...)
	(... Then:  "Target Architecture (PowerPC64 (little endian)) --->"...)
	(... Select: "(X) PowerPC64 (big endian)")
	(... Save and exit!... )

Finally, build the toolchain and grab a coffee:

	$ make sdk
	(... drink the coffee and don't self burn...)

In the end, install the toolchain to the host system (in /opt) and
relocate it it:

	$ sudo tar xf output-ppc64/images/powerpc64-buildroot-linux-gnu_sdk-buildroot.tar.gz -C /opt
	$ sudo /opt/powerpc64-buildroot-linux-gnu_sdk-buildroot/relocate-sdk.sh /opt/powerpc64-buildroot-linux-gnu_sdk-buildroot/

Update the PATH variable with the location of the toolchain and re-login
to the user session:

	$ sudo printf "#/bin/sh\n\nexport PATH=\"\$PATH:/opt/powerpc64-buildroot-linux-gnu_sdk-buildroot/bin\"\n" | tee /etc/profile.d/powerpc64-buildroot-linux-gnu_sdk-buildroot.sh

Or source the profile script in every terminal to run the toolchain
binaries without its absolute path:

	$ source /etc/profile.d/powerpc64-buildroot-linux-gnu_sdk-buildroot.sh

Run:

	$ make powerpc64-archlinuxpower-chroot
	(...)
	[root@archlinux /]# uname -a
	Linux archlinux 6.7.3-arch1-2 #1 SMP PREEMPT_DYNAMIC Fri, 02 Feb 2024 17:03:55 +0000 ppc64 GNU/Linux

Note: In the ArchLinuxPOWER PowerPC64 LE, the result is:

	$ make powerpc64le-archlinuxpower-chroot
	(...)
	[root@archlinux /]# uname -a
	Linux archlinux 6.7.3-arch1-2 #1 SMP PREEMPT_DYNAMIC Fri, 02 Feb 2024 17:03:55 +0000 ppc64le GNU/Linux

[1]: https://git.musl-libc.org/cgit/musl/commit/?id=c0ede9e4046a0882d83ae4b45c7dfac86fb7c15d
[2]: https://git.musl-libc.org/cgit/musl/tree/configure?id=c0ede9e4046a0882d83ae4b45c7dfac86fb7c15d#n649
[3]: https://toolchains.bootlin.com/
[4]: https://github.com/bootlin/buildroot-toolchains
[5]: https://github.com/buildroot/buildroot/commit/5ab751ca44e9da6f406876ff582b3a813056b0d4
[6]: https://archlinuxpower.org/
[7]: https://toolchains.bootlin.com/releases_powerpc64-power8.html