{get,set}groups: IAMROOT_GROUPS use whitespace delimiter
According to bash(1):

	GROUPS An array variable containing the list of groups of which
	the current user is a member. Assignments to GROUPS have no
	effect. If GROUPS is unset, it loses its special properties,
	even if it is subsequently reset.

The variable GROUPS is a bash array variable, therefore it is separated
by IFS.

	IFS The Internal Field Separator that is used for word splitting
	after expansion and to split lines into words with the read
	builtin command. The default value is ‘‘<space><tab><newline>''.

The variable IAMROOT_GROUPS is similar to GROUPS but it is an
environment variable as it cannot be an array.

The bash arrays are expanded using the first character of the IFS
special variable. If IFS is unset, the array members are separated by
spaces.

As a consequence, and unlike the bash variable GROUPS, the environment
variable IAMROOT_GROUPS uses the colon delimiter instead of "space"
(i.e. the default if IFS is unset).

This updates the delimiter used by environment variable IAMROOT_GROUPS
to whitespace to intend to get closer to the bash variable GROUPS, even
if is not possible.