Fix "Add IAMROOT_PATH_RESOLUTION_AF_UNIX"
This fixes commit 8a7b092984f095870c20d4396482b1e357a75385.

The environment variable IAMROOT_PATH_RESOLUTION_AF_UNIX was introduced
to reduce the path resolution of an Unix socket if it exceeds the 108
characters.

However, the feature is broken since its introduction as it prepends the
already too-long (and truncated) resolved path instead of the filename
of the original path; the new resolved path is "somehow" even longer.

This fixes the new path resolution to prepend the variable environment
IAMROOT_PATH_RESOLUTION_AF_UNIX to the basename of the original Unix
socket path.