Test chroot'ing to an s390x's musl world
The project uses the pre-built toolchains provided by Bootlin[1] since
commit b00046e4bbf1a6849f081a5c398d16be0b52cda3 to compile the libraries
targetting a foreign world.

Bootlin uses Buildroot to make its toolchains[2].

The package musl in Buildroot does not support s390x[3] even though musl
does support it since a while (1.1.16, 2016)[4].

As a consequence, there is no s390x musl Bootlin toolchain.

Follow the commands below to build the toolchain and to install it to the
host system.

First, clone the repository and patch the sources with the patch[5]:

	$ git clone https://gitlab.com/buildroot.org/buildroot.git
	$ cd buildroot

Then, configure using QEMU PowerPC64 LE (POWER8) and update the target
architecture to PowerPC64 (big endian):

	$ ( cd configs && curl -O https://github.com/gportay/iamroot/blob/master/support/toolchain_alpinelinux_s390x_defconfig )
    	$ make toolchain_alpinelinux_s390x_defconfig

Finally, build the toolchain and grab a coffee:

	$ make sdk
	(... drink the coffee and don't self burn...)

In the end, install the toolchain to the host system (in /opt) and
relocate it:

    	$ sudo tar xf output/images/s390x-buildroot-linux-musl_sdk-buildroot.tar.gz -C /opt
    	$ sudo /opt/s390x-buildroot-linux-musl_sdk-buildroot.tar.gz/relocate-sdk.sh /opt/s390x-buildroot-linux-musl_sdk-buildroot/

Update the PATH variable with the location of the toolchain and re-login
to the user session:

    	$ sudo printf "#/bin/sh\n\nexport PATH=\"\$PATH:/opt/s390x-buildroot-linux-musl_sdk-buildroot/bin\"\n" | tee /etc/profile.d/s390x-buildroot-linux-musl_sdk-buildroot.sh

Or source the profile script in every terminal to run the toolchain
binaries without its absolute path:

	$ source /etc/profile.d/s390x-buildroot-linux-musl_sdk-buildroot.sh

Run:
	$ make s390x-alpinelinux-edge-chroot
	(...)
	/ # uname -a
	Linux archlinux 6.7.4-arch1-1 #1 SMP PREEMPT_DYNAMIC Mon, 05 Feb 2024 22:07:49 +0000 s390x Linux
	/ # /lib/ld-musl-s390x.so.1
	musl libc (s390x)
	Version 1.2.4_git20230717
	Dynamic Program Loader
	Usage: /lib/ld-musl-s390x.so.1 [options] [--] pathname [args]

[1]: https://toolchains.bootlin.com/
[2]: https://github.com/bootlin/buildroot-toolchains
[3]: https://gitlab.com/buildroot.org/buildroot/-/blob/2023.11.x/package/musl/Config.in?ref_type=heads#L3-L20
[4]: https://git.musl-libc.org/cgit/musl/commit/?id=15094943050eb9a564f409323070e50b40f78816
[5]: https://patchwork.ozlabs.org/project/buildroot/patch/20240218011118.1077002-1-gael.portay@rtone.fr/