REWORD: dso: decrease massively file-descriptor open usage
The library uses massively the environ mostly for setup and additionaly
for dynamic allocation (aka. temporary preload, library-path, argv0,
inhibit-rpath for the functions execve{,at}(), posix_spawn(), and
dlopen()).

The later is unnecessary and it can be replaced by allocating huge
buffers on the stack.

This decreases slightly the amount of environ variable used and removes
the variables ld_preload, ld_library_path and inhibit_rpath (argv0
remains for dynamic loaders not supporting --argv0).

Note: This removes the function __path_setenv().