execve{,at}, posix_spawn: preload library if not chroot'ed
The functions execve(), execveat() and posix_spawn() do not proceed to
any hack if not chroot'ed. They call the original directly if the
executable is not ignored or is not SUID.

As a consequence, the command-line is not rewritten and the executable
may run out of the iamroot environment if LD_PRELOAD is unset.

In the linux world (musl and glibc with support for option --preload),
the variables LD_PRELOAD and LD_LIBRARY_PATH are cleared since commit
5201fdd3d981f1de2d32c19778610b385038484c.

Therefore, the iamroot environment is lost for the processes enterring
chroot (unsetenv LD_PRELOAD) and exiting chroot then (empty LD_PRELOAD
and IAMROOT_ROOT).

RPM has introduced new way to create user and group since 4.19[1][2];
the new shell scriptlet sysusers.sh[3] is not run in the chroot'ed
environment. Intstead, it runs the useradd and groupadd using the option
-R to chroot in the rootfs.

This ensure to preload the iamroot library if LD_PRELOAD is unset and if
not in a chroot'ed environment.

[1]: https://github.com/rpm-software-management/rpm/blob/master/docs/manual/users_and_groups.md
[2]: https://github.com/rpm-software-management/rpm/commit/2bdd7d7cbd77b13abf512bb919abe7f0537a2c45
[3]: https://github.com/rpm-software-management/rpm/blob/rpm-4.19.1.1-release/scripts/sysusers.sh#L122