dso: fix circular dependencies in __ldso_preload_needed()
The function __ldso_preload_needed() appends the needed shared-objects
first and the shared-object then.

As a consequence, if a needed shared-object of the shared-object needs
the shared-object itself (i.e. a circular dependency), the function
__ldso_preload_needed() via __path_iterate() loops forever till the
stack is full and raises a segmentation fault.

Mainly because the context lacks for the list of the shared-objects
actually parsed.

This fixes the circular dependency by updating the preload processing
order of shared-objects: the shared-object is now appended to the buffer
before its needed shared-objects.

Therefore, the callback __ldso_preload_needed_callback() skips the
shared-object actually parsed and does the go into an infinite loop.

Fixes:

	[root@archlinux /]# pacman -S sway
	(...)
	(8/8) Updating fontconfig cache...
	error: command terminated by signal 11: Segmentation fault

	[root@archlinux /]# fc-cache -rs
	Segmentation fault (core dumped)