iamroot: remove detailled execve output command on level 4 and above
The exec() family of functions replaces the current process image with a
new process image.

As a consequence, it is impossible to output the execve traces after the
function is actually called; the traces are output at the begining of
the intercepted function, without the returned value.

The function __verbose_exec() is a second trace that outputs the argv
and envp of the (possible) rewritten command-line, right before running
the true function.

Note: The function __verbose_exec() is misnamed as the detailled view
outputs the command-line as execve() and it is used by the execveat()
since commit 11dec911c9cbe3e1b9652daac591383388f6c753 and posix_spawn()
since commit 664bce67e7e2d463cc9d8ce33af6fcceb2d305c0.

The detailled ouput become pointless since the full command is now
outputted since the commit 1f2f0a7bf23c411cba05affecbcd302577d8bada.

For all this reasons, this removes the detailled output entirelly in
favor of the full command starting at level 1.