dso: compare machine systemd and co against 1 in if statements
The functions __is_arm() and __is_armhf() returns -1 and ENOTSUP since
commit 9dd7816a78a2aa1d8cae3d167e486773f5b184d5 if the EM_ machine macro
is undefined (mostly because *BSD systems do not defined them if they do
not support the machines).

This is generatlized later to all the supported machine functions (i.e.
__is_aarch64(), __is_aarch64_be(), __is_riscv(), __is_mipsle(),
__is_powerpc64(), __is_powerpc64le() and __is_s390()) since the commit
36355628cccc6ddce235d19fab916301a4e4e33a

The tests must be updated to compare the returned value against 1 to
ensure it is the expected result (if (-1) is true).

This fixes commit 9dd7816a78a2aa1d8cae3d167e486773f5b184d5.