dso: ensure the path to binary is set in __ldso()
The dynamic loaders on NetBSD and OpenBSD cannot be run directly from
the command-line; the former causes a segmentation fault, while the
other does not have the execute bit set.

The function __ldso() does nothing to interparg if NetBSD or OpenBSD and
returns 0.

The functions execve(), execveat() and posix_spawn() saves argv[0] to
the interparg[0] and forwards interparg[0] as the program to execute to
the intercepted function.

As a consequence, the function fails if argv[0] with ENOENT if it is not
an absolute path (i.e. the filename only) or if argv[0] refers to a
login shell (i.e. prepend with a dash, as in -sh).

This ensures the function __ldso() sets the resolved path for the
program and NULL at least (and return 1), in order to overwrite the
previous value (i.e. argv[0]).

This fixes commit e52255a819ed3b7195de8208c941b7f22fa76e43.

Fixes (on *BSD):

	-bash-5.2$ make test
	(...)
	cd rootfs/ && PATH=/home/freebox/src/iamroot/tests/rootfs/usr/bin env LD_PRELOAD=/home/freebox/src/iamroot/aarch64/libiamroot.elf_so:/home/freebox/src/iamroot/tests/rootfs/lib/libc.so.12 LD_LIBRARY_PATH=/home/freebox/src/iamroot/tests/rootfs/usr/lib IAMROOT_ROOT=/home/freebox/src/iamroot/tests/rootfs PWD=/ test-execl                                                                      | tee /dev/stderr | grep '^one two three$'
	execl: No such file or directory