dso: do not preload iamroot library if preloading already
The iamroot library preloads itself if the new process loses the preload.

The internal function __is_preloading_so() checks if the iamroot library
is preloaded via the environment variable LD_PRELOAD. It uses hard-coded
values and the content of IAMROOT_LIB.

The function compares absolute paths. As a consequence, it fails to
check for already preloading if the iamroot library is preloaded from a
location that is different from the hard-coded values.

Then, the iamroot library is preloaded more than once and it causes the
process to crash with a segmentation fault as multiple preloading is not
supported (yet?).

This fixes the segmentation fault if the library preloads itself more
than once by checking filename instead of absolute path.