dso: fix setting invalid IAMROOT_LIB in __getlibiamroot()
The function __getlibiamroot() may set the environment variable
IAMROOT_LIB using the value from the environment.

As a consequence, setenv() may update the __environ; the call may
realloc() the __environ, and thus the string pointed by the variable is
invalid. The string memory is free'd and it may be alterned by another
thread in the meanwhile.

This sets the environment variable IAMROOT_LIB after the upcoming value
is copied to the persistent buffer (it is stack allocated) to avoid an
invalid write caused by an invalid read.