statx,fstat{,at}: decrease slightly extended attribute usage
The library uses __get_uid() and __get_gid() twice: once to set the
st_uid and st_gid to 0 if user iamroot.uid and group iamroot.gid are
unset; then to set st_uid and st_gid if user iamroot.uid and group
iamroot.gid are set through the macros __st_uid() and __st_gid().

The former can be merged to the later to avoid having two calls.

This decreases slightly the amount of extend attribute usage by moving
variable used and removes the variables .