linux.mk: fix remake of adelielinux-rootfs target
Unlike Alpine Linux, Adelie Linux is based on the GNU coreutils; i.e. it
is not based on busybox.

The adelielinux-rootfs target pre-requires the file -rootfs/bin/busybox,
that do not exist in the Adelie Linux world, thus the target is remake.

This fixes the remake of adelielinux-rootfs target by pre-requiring the
the file -rootfs/bin/sh as every other GNU coreutils distributions.