Do not follow symlinks in __path_access()
The parameter flag in function faccessat() supports the value AT_EACCESS
only; the value AT_SYMLINK_NOFOLLOW is not specified.

The man-page access(3p) is unclear:

	DESCRIPTION

	(...)

	Values for flag are constructed by a bitwise-inclusive OR of
	flags from the following list, defined in <fcntl.h>:

	AT_EACCESS The checks for accessibility (including directory
		   permissions checked during pathname resolution) shall
		   be performed using the effective user ID and group ID
		   instead of the real user ID and group ID as required
	           in a call to access().

	(...)

	ERRORS

	(...)

	The faccessat() function may fail if:

	EINVAL The value of the flag argument is not valid.

The function call faccessat(AT_FDCWD, path, F_OK, AT_SYMLINK_NOFOLLOW)
behaves differently in a GNU and in musl world on Linux. The function
call returns 0 in a GNU world[1], and it returns -1/EINVAL in a musl
world[2].

The GNU C Library supports AT_SYMLINK_NOFOLLOW while musl does not; see
threads[3][4] and the workaround[5][6] then the fix [7][8] on toybox.

This introduces the function __faccessat() that supports the AT-flag
AT_SYMLINK_NOFOLLOW for internal purpose. It is used then to avoid
following symlinks in the function __path_access().

Fixes:

	Running ld-iamroot.so: test option -R /home/gportay/src/iamroot/tests/rootfs changes root directory...
	execvpe: No such file or directory
	ld-iamroot.so: test option -R /home/gportay/src/iamroot/tests/rootfs changes root directory: [KO]

[1]: https://sourceware.org/git/?p=glibc.git;a=blob;f=sysdeps/unix/sysv/linux/faccessat.c;h=80b3b240f7ca05d36d4da451ee83245893077b0b#l37
[2]: https://git.musl-libc.org/cgit/musl/commit/src/unistd/faccessat.c?id=0a05eace163cee9b08571d2ff9d90f5e82d9c228
[3]: https://www.openwall.com/lists/musl/2014/09/25/1
[4]: https://www.openwall.com/lists/musl/2015/02/04/5
[5]: https://github.com/landley/toybox/commit/2ccab0260f6812f78cc0bafdf2bb2a92942ee6f3
[6]: https://github.com/landley/toybox/blob/037009f8f65b73264d2f131faa4efa36b11a27d0/www/news.html#L115-L125
[7]: https://github.com/landley/toybox/commit/bd6c3f35ae226125737503dddac48e7924425b49
[8]: https://github.com/landley/toybox/blob/ba3e32b81935135f8560f59e44d87d032fbc4653/www/news.html#L125-L126