ish: add --login option
According to bash(1):

	When bash is invoked as an interactive login shell, or as a
	non-interactive shell with the --login option, it first reads
	and executes commands from the file /etc/profile, if that file
	exists. After reading that file, it looks for ~/.bash_profile,
	~/.bash_login, and ~/.profile, in that order, and reads and
	executes commands from the first one that exists and is
	readable. The --noprofile option may be used when the shell is
	started to inhibit this behavior.

	When an interactive login shell exits, or a non-interactive
	login shell executes the exit builtin command, bash reads and
	executes commands from the file ~/.bash_logout, if it exists.

	When an interactive shell that is not a login shell is started,
	bash reads and executes commands from ~/.bashrc, if that file
	exists. This may be inhibited by using the --norc option. The
	--rcfile file option will force bash to read and execute
	commands from file instead of ~/.bashrc.

This adds option --login to act as if it had been invoked as a login
shell, to read and execute commands from the files $HOME/.ish_profile
and $PWD/.ishrc if these files exist.

Note: The login shell does **NOT** make the difference between a login
shell and an interactive shell; it does **NOT** read and executes a
system-wide startup file, personal initialization files or an exit file.
Instead, it reads and executes a user-wide startup file and a single
location initialization file. The environment variables ISH_PROFILE_FILE
and ISH_RC_FILE may be set to empty before the shell is started to
inhibit this behavior.